Move resolve func to DNSNodeRecord::
