+++ bin/reth/src/commands/p2p/mod.rs
@@ -96,8 +96,8 @@ impl Command {
 
         let mut config: Config = confy::load_path(&config_path).unwrap_or_default();
 
-        for &peer in &self.network.trusted_peers {
-            config.peers.trusted_nodes.insert(peer);
+        for peer in &self.network.trusted_peers {
+            config.peers.trusted_nodes.insert(peer.resolve(None).await?);
         }
 
         if config.peers.trusted_nodes.is_empty() && self.network.trusted_only {
