+++ crates/net/common/Cargo.toml
@@ -12,14 +12,9 @@ description = "Types shared across network code"
 workspace = true
 
 [dependencies]
-# reth
-reth-network-types.workspace = true
-
 # ethereum
 alloy-primitives = { workspace = true, features = ["rand"] }
 
 # async
 pin-project.workspace = true
-url.workspace = true
 tokio = { workspace = true, features = ["full"] }
-tokio-retry = { workspace = true }
