+++ crates/net/common/src/lib.rs
@@ -17,6 +17,4 @@ pub mod bandwidth_meter;
 /// Traits related to tokio streams
 pub mod stream;
 
-pub mod dns_node_record_resolve;
-
 pub mod ratelimit;
