+++ crates/net/types/Cargo.toml
@@ -25,6 +25,8 @@ secp256k1.workspace = true
 serde_with.workspace = true
 thiserror.workspace = true
 url.workspace = true
+tokio = { workspace = true, features = ["full"] }
+tokio-retry = { workspace = true }
 
 [dev-dependencies]
 alloy-primitives = { workspace = true, features = ["rand"] }
