+++ crates/net/types/src/dns_node_record.rs
@@ -2,6 +2,7 @@
 
 use std::{
     fmt::{self, Write},
+    io::Error,
     net::IpAddr,
     num::ParseIntError,
     str::FromStr,
@@ -27,6 +28,15 @@ pub struct DNSNodeRecord {
     pub id: PeerId,
 }
 
+/// Retry strategy for DNS lookups.
+#[derive(Debug)]
+pub struct RetryStrategy {
+    /// The amount of time between retries.
+    interval: u64,
+    /// The number of retries.
+    retries: usize,
+}
+
 impl DNSNodeRecord {
     /// Derive the [`NodeRecord`] from the secret key and addr
     pub fn from_secret_key(host: Host, port: u16, sk: &SecretKey) -> Self {
@@ -38,6 +48,65 @@ impl DNSNodeRecord {
     /// Creates a new record from a socket addr and peer id.
     pub fn new(host: Host, port: u16, id: PeerId) -> Self {
         Self { host, tcp_port: port, udp_port: port, id }
+    }
+
+    /// Resolves the host in a [DNSNodeRecord] to an IP address, returning a [NodeRecord].
+    pub async fn resolve(
+        &self,
+        retry_strategy: Option<RetryStrategy>,
+    ) -> Result<NodeRecord, Error> {
+        let domain = match self.host.to_owned() {
+            Host::Ipv4(ip) => {
+                let id = self.id;
+                let tcp_port = self.tcp_port;
+                let udp_port = self.udp_port;
+
+                return Ok(NodeRecord { address: ip.into(), id, tcp_port, udp_port })
+            }
+            Host::Ipv6(ip) => {
+                let id = self.id;
+                let tcp_port = self.tcp_port;
+                let udp_port = self.udp_port;
+
+                return Ok(NodeRecord { address: ip.into(), id, tcp_port, udp_port })
+            }
+            Host::Domain(domain) => domain,
+        };
+
+        // Lookip ipaddr from domain
+        match Self::lookup_host(&domain).await {
+            Ok(ip) => Ok(NodeRecord {
+                address: ip,
+                id: self.id,
+                tcp_port: self.tcp_port,
+                udp_port: self.udp_port,
+            }),
+            Err(e) => {
+                // Retry if strategy is provided
+                if let Some(strat) = retry_strategy {
+                    let lookup = || async { Self::lookup_host(&domain).await };
+                    let strat = tokio_retry::strategy::FixedInterval::from_millis(strat.interval)
+                        .take(strat.retries);
+                    let ip = tokio_retry::Retry::spawn(strat, lookup).await?;
+                    Ok(NodeRecord {
+                        address: ip,
+                        id: self.id,
+                        tcp_port: self.tcp_port,
+                        udp_port: self.udp_port,
+                    })
+                } else {
+                    Err(e)
+                }
+            }
+        }
+    }
+
+    async fn lookup_host(domain: &str) -> Result<std::net::IpAddr, Error> {
+        let mut ips = tokio::net::lookup_host(format!("{domain}:0")).await?;
+        let ip = ips
+            .next()
+            .ok_or_else(|| Error::new(std::io::ErrorKind::AddrNotAvailable, "No IP found"))?;
+        Ok(ip.ip())
     }
 }
 
@@ -232,6 +301,32 @@ mod tests {
         for (url, expected) in cases {
             let node: DNSNodeRecord = serde_json::from_str(url).expect("couldn't deserialize");
             assert_eq!(node, expected);
+        }
+    }
+
+    #[tokio::test]
+    async fn test_resolve_dns_node_record() {
+        use super::*;
+
+        // Set up tests
+        let tests = vec![
+            ("localhost", None),
+            ("localhost", Some(RetryStrategy { interval: 1000, retries: 3 })),
+        ];
+
+        // Run tests
+        for (domain, retry_strategy) in tests {
+            // Construct record
+            let rec =
+                DNSNodeRecord::new(url::Host::Domain(domain.to_owned()), 30300, PeerId::random());
+
+            // Resolve domain and validate
+            let rec = rec.resolve(retry_strategy).await.unwrap();
+            if let std::net::IpAddr::V4(addr) = rec.address {
+                assert_eq!(addr, std::net::Ipv4Addr::new(127, 0, 0, 1))
+            } else {
+                panic!("Expected IPv4 address");
+            }
         }
     }
 }
