+++ crates/node/builder/Cargo.toml
@@ -26,7 +26,6 @@ reth-rpc-layer.workspace = true
 reth-node-api.workspace = true
 reth-node-core.workspace = true
 reth-network.workspace = true
-reth-net-common.workspace = true
 reth-primitives.workspace = true
 reth-payload-builder.workspace = true
 reth-transaction-pool.workspace = true
